edited fms and distributor

+++ src/context/DistributorContext.js
@@ -1,6 +1,8 @@
 import { createContext, useState } from "react";
 
 import axios from "axios";
+import { Navigate } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const DistributorContext = createContext();
 
@@ -24,6 +26,13 @@ export function DistributorContextProvider({ children }) {
 
     setDetailData(distributorObject.data);
   }
+  async function returnDistributorById(id) {
+    const distributorObject = await axios.get(
+      `https://localhost:44393/api/GetDirstributorById/${id}`
+    );
+
+    return distributorObject;
+  }
 
   async function deleteDistributor(id) {
     const deleteddistributor = await axios.delete(
@@ -31,10 +40,40 @@ export function DistributorContextProvider({ children }) {
     );
   }
 
-  const handleDelete = (id) => {
-    deleteDistributor(id);
+  async function getAlldistdel() {
+    const alldist = await axios.get(
+      `https://localhost:44393/api/GetAllDistributors`
+    );
+    return alldist.data;
+  }
 
-    getAlldistributor();
+  const handleDelete = (id) => {
+    Swal.fire({
+      icon: "warning",
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "No, cancel!",
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire({
+          icon: "success",
+          title: "Deleted!",
+          showConfirmButton: false,
+          timer: 2000,
+        });
+
+        deleteDistributor(id);
+        setData(getAlldistdel());
+        setData(data.filter((employee) => employee.distributorId !== id));
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        Swal.fire("Cancelled", "Your imaginary file is safe :)", "error");
+      }
+    });
   };
 
   async function updateDistributor(id, updatedData) {
@@ -52,7 +91,6 @@ export function DistributorContextProvider({ children }) {
 
   const handleupdate = (id, updatedData) => {
     updateDistributor(id, updatedData);
-
     getAlldistributor();
   };
 
@@ -65,6 +103,7 @@ export function DistributorContextProvider({ children }) {
     handleupdate,
     getDistributorById,
     getDistributorProduct,
+    returnDistributorById,
   };
   return (
     <DistributorContext.Provider value={valuetoshare}>
