+++ src/context/StatementContext.js
@@ -1,6 +1,7 @@
 import { createContext, useState } from "react";
 
 import axios from "axios";
+import Swal from "sweetalert2";
 
 const StatementContext = createContext();
 
@@ -37,10 +38,39 @@ export function StatementContextProvider({ children }) {
     );
   }
 
-  const handleDelete = (id) => {
-    deleteStatement(id);
+  async function getAlldistdel() {
+    const alldist = await axios.get(
+      `https://localhost:44393/api/FmsGetAllStatements`
+    );
+    return alldist.data;
+  }
 
-    getAllstatement();
+  const handleDelete = (id) => {
+    Swal.fire({
+      icon: "warning",
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "No, cancel!",
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire({
+          icon: "success",
+          title: "Deleted!",
+          showConfirmButton: false,
+          timer: 2000,
+        });
+
+        deleteStatement(id);
+        setData(getAlldistdel());
+        setData(data.filter((employee) => employee.staId !== id));
+      } else if (
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        Swal.fire("Cancelled", "Your imaginary file is safe :)", "error");
+      }
+    });
   };
 
   async function updateStatement(id, updatedData) {
@@ -48,6 +78,13 @@ export function StatementContextProvider({ children }) {
       `https://localhost:44393/api/FmsUpdateStatement/${id}`,
       updatedData
     );
+  }
+  async function returnStatementById(id) {
+    const statementObject = await axios.get(
+      `https://localhost:44393/api/FmsGetStatementById/${id}`
+    );
+
+    return statementObject;
   }
 
   const handleupdate = (id, updatedData) => {
@@ -64,6 +101,7 @@ export function StatementContextProvider({ children }) {
     handleDelete,
     handleupdate,
     getStatementById,
+    returnStatementById,
   };
   return (
     <StatementContext.Provider value={valuetoshare}>
